+++ deploy/Chart/charts/de/templates/serviceaccount.yaml
@@ -2,40 +2,40 @@ apiVersion: v1
 kind: ServiceAccount
 metadata:
   name: de-manager
-  namespace: {{ .Release.Namespace }}
+  namespace: { { .Release.Namespace } }
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
   name: de-manager-role
 rules:
-- apiGroups:
-  - ""
-  resources:
-  - configmaps
-  - secrets
-  - services
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - apps
-  resources:
-  - deployments
-  - statefulsets
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
+  - apiGroups:
+      - ""
+    resources:
+      - configmaps
+      - secrets
+      - services
+    verbs:
+      - create
+      - delete
+      - get
+      - list
+      - patch
+      - update
+      - watch
+  - apiGroups:
+      - apps
+    resources:
+      - deployments
+      - statefulsets
+    verbs:
+      - create
+      - delete
+      - get
+      - list
+      - patch
+      - update
+      - watch
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
@@ -46,6 +46,6 @@ roleRef:
   kind: ClusterRole
   name: de-manager-role
 subjects:
-- kind: ServiceAccount
-  name: de-manager
-  namespace: {{ .Release.Namespace }}
+  - kind: ServiceAccount
+    name: de-manager
+    namespace: { { .Release.Namespace } }
