+++ pkg/cli/environments/azure.go
@@ -30,22 +30,9 @@ func RequireAzureCloud(e Environment) (*AzureCloudEnvironment, error) {
 
 // AzureCloudEnvironment represents an Azure Cloud Radius environment.
 type AzureCloudEnvironment struct {
-	Name                     string `mapstructure:"name" validate:"required"`
-	Kind                     string `mapstructure:"kind" validate:"required"`
-	SubscriptionID           string `mapstructure:"subscriptionid" validate:"required"`
-	ResourceGroup            string `mapstructure:"resourcegroup" validate:"required"`
-	ClusterName              string `mapstructure:"clustername" validate:"required"`
-	DefaultApplication       string `mapstructure:"defaultapplication" yaml:",omitempty"`
-	Context                  string `mapstructure:"context" validate:"required"`
-	Namespace                string `mapstructure:"namespace" validate:"required"`
-	RadiusRPLocalURL         string `mapstructure:"radiusrplocalurl,omitempty"`
-	DeploymentEngineLocalURL string `mapstructure:"deploymentenginelocalurl,omitempty"`
-	UCPLocalURL              string `mapstructure:"ucplocalurl,omitempty"`
-	EnableUCP                bool   `mapstructure:"enableucp,omitempty"`
-	UCPResourceGroupName     string `mapstructure:"ucpresourcegroupname,omitempty"`
-
-	// We tolerate and allow extra fields - this helps with forwards compat.
-	Properties map[string]interface{} `mapstructure:",remain" yaml:",omitempty"`
+	RadiusEnvironment
+	SubscriptionID string `mapstructure:"subscriptionid" validate:"required"`
+	ResourceGroup  string `mapstructure:"resourcegroup" validate:"required"`
 }
 
 func (e *AzureCloudEnvironment) GetName() string {
@@ -61,7 +48,7 @@ func (e *AzureCloudEnvironment) GetDefaultApplication() string {
 }
 
 func (e *AzureCloudEnvironment) GetKubeContext() string {
-	return e.Context
+	return e.KubeContext
 }
 
 func (e *AzureCloudEnvironment) GetContainerRegistry() *Registry {
@@ -82,7 +69,7 @@ func (e *AzureCloudEnvironment) CreateDeploymentClient(ctx context.Context) (cli
 	url, roundTripper, err := kubernetes.GetBaseUrlAndRoundTripperForDeploymentEngine(
 		e.DeploymentEngineLocalURL,
 		e.UCPLocalURL,
-		e.Context,
+		e.KubeContext,
 		e.EnableUCP,
 	)
 	if err != nil {
@@ -128,17 +115,17 @@ func (e *AzureCloudEnvironment) CreateDeploymentClient(ctx context.Context) (cli
 }
 
 func (e *AzureCloudEnvironment) CreateDiagnosticsClient(ctx context.Context) (clients.DiagnosticsClient, error) {
-	k8sTypedClient, config, err := kubernetes.CreateTypedClient(e.Context)
+	k8sTypedClient, config, err := kubernetes.CreateTypedClient(e.KubeContext)
 	if err != nil {
 		return nil, err
 	}
 
-	k8sRuntimeclient, err := kubernetes.CreateRuntimeClient(e.Context, kubernetes.Scheme)
+	k8sRuntimeclient, err := kubernetes.CreateRuntimeClient(e.KubeContext, kubernetes.Scheme)
 	if err != nil {
 		return nil, err
 	}
 
-	_, con, err := kubernetes.CreateAPIServerConnection(e.Context, e.RadiusRPLocalURL)
+	_, con, err := kubernetes.CreateAPIServerConnection(e.KubeContext, e.RadiusRPLocalURL)
 	if err != nil {
 		return nil, err
 	}
@@ -154,7 +141,7 @@ func (e *AzureCloudEnvironment) CreateDiagnosticsClient(ctx context.Context) (cl
 }
 
 func (e *AzureCloudEnvironment) CreateLegacyManagementClient(ctx context.Context) (clients.LegacyManagementClient, error) {
-	_, connection, err := kubernetes.CreateAPIServerConnection(e.Context, e.RadiusRPLocalURL)
+	_, connection, err := kubernetes.CreateAPIServerConnection(e.KubeContext, e.RadiusRPLocalURL)
 	if err != nil {
 		return nil, err
 	}
@@ -168,7 +155,7 @@ func (e *AzureCloudEnvironment) CreateLegacyManagementClient(ctx context.Context
 }
 
 func (e *AzureCloudEnvironment) CreateApplicationsManagementClient(ctx context.Context) (clients.ApplicationsManagementClient, error) {
-	_, connection, err := kubernetes.CreateAPIServerConnection(e.Context, e.UCPLocalURL)
+	_, connection, err := kubernetes.CreateAPIServerConnection(e.KubeContext, e.UCPLocalURL)
 	if err != nil {
 		return nil, err
 	}
