+++ pkg/cli/environments/kubernetes.go
@@ -22,19 +22,7 @@ import (
 
 // KubernetesEnvironment represents a Kubernetes Radius environment.
 type KubernetesEnvironment struct {
-	Name                     string `mapstructure:"name" validate:"required"`
-	Kind                     string `mapstructure:"kind" validate:"required"`
-	Context                  string `mapstructure:"context" validate:"required"`
-	Namespace                string `mapstructure:"namespace" validate:"required"`
-	DefaultApplication       string `mapstructure:"defaultapplication,omitempty"`
-	RadiusRPLocalURL         string `mapstructure:"radiusrplocalurl,omitempty"`
-	DeploymentEngineLocalURL string `mapstructure:"deploymentenginelocalurl,omitempty"`
-	UCPLocalURL              string `mapstructure:"ucplocalurl,omitempty"`
-	EnableUCP                bool   `mapstructure:"enableucp,omitempty"`
-	UCPResourceGroupName     string `mapstructure:"ucpresourcegroupname,omitempty"`
-
-	// We tolerate and allow extra fields - this helps with forwards compat.
-	Properties map[string]interface{} `mapstructure:",remain"`
+	RadiusEnvironment
 }
 
 func (e *KubernetesEnvironment) GetName() string {
@@ -54,7 +42,7 @@ func (e *KubernetesEnvironment) GetDefaultApplication() string {
 }
 
 func (e *KubernetesEnvironment) GetKubeContext() string {
-	return e.Context
+	return e.KubeContext
 }
 
 func (e *KubernetesEnvironment) GetContainerRegistry() *Registry {
@@ -77,7 +65,7 @@ func (s *sender) Do(request *http.Request) (*http.Response, error) {
 }
 
 func (e *KubernetesEnvironment) CreateDeploymentClient(ctx context.Context) (clients.DeploymentClient, error) {
-	url, roundTripper, err := kubernetes.GetBaseUrlAndRoundTripperForDeploymentEngine(e.DeploymentEngineLocalURL, e.UCPLocalURL, e.Context, e.EnableUCP)
+	url, roundTripper, err := kubernetes.GetBaseUrlAndRoundTripperForDeploymentEngine(e.DeploymentEngineLocalURL, e.UCPLocalURL, e.KubeContext, e.EnableUCP)
 
 	if err != nil {
 		return nil, err
@@ -103,16 +91,16 @@ func (e *KubernetesEnvironment) CreateDeploymentClient(ctx context.Context) (cli
 }
 
 func (e *KubernetesEnvironment) CreateDiagnosticsClient(ctx context.Context) (clients.DiagnosticsClient, error) {
-	k8sClient, config, err := kubernetes.CreateTypedClient(e.Context)
+	k8sClient, config, err := kubernetes.CreateTypedClient(e.KubeContext)
 	if err != nil {
 		return nil, err
 	}
-	client, err := kubernetes.CreateRuntimeClient(e.Context, kubernetes.Scheme)
+	client, err := kubernetes.CreateRuntimeClient(e.KubeContext, kubernetes.Scheme)
 	if err != nil {
 		return nil, err
 	}
 
-	_, con, err := kubernetes.CreateAPIServerConnection(e.Context, e.RadiusRPLocalURL)
+	_, con, err := kubernetes.CreateAPIServerConnection(e.KubeContext, e.RadiusRPLocalURL)
 	if err != nil {
 		return nil, err
 	}
@@ -128,7 +116,7 @@ func (e *KubernetesEnvironment) CreateDiagnosticsClient(ctx context.Context) (cl
 }
 
 func (e *KubernetesEnvironment) CreateLegacyManagementClient(ctx context.Context) (clients.LegacyManagementClient, error) {
-	_, connection, err := kubernetes.CreateAPIServerConnection(e.Context, e.RadiusRPLocalURL)
+	_, connection, err := kubernetes.CreateAPIServerConnection(e.KubeContext, e.RadiusRPLocalURL)
 	if err != nil {
 		return nil, err
 	}
@@ -142,7 +130,7 @@ func (e *KubernetesEnvironment) CreateLegacyManagementClient(ctx context.Context
 }
 
 func (e *KubernetesEnvironment) CreateApplicationsManagementClient(ctx context.Context) (clients.ApplicationsManagementClient, error) {
-	_, connection, err := kubernetes.CreateAPIServerConnection(e.Context, e.UCPLocalURL)
+	_, connection, err := kubernetes.CreateAPIServerConnection(e.KubeContext, e.UCPLocalURL)
 	if err != nil {
 		return nil, err
 	}
